fix(schedule): expose other fields when id optional (#451)
This PR appears to follow-up on
https://github.com/apify/apify-client-js/pull/276 which made it so `id`
was optional ✅ but did so in a way that removed the remainder of the
required fields from the type definition ❌, e.g. `actorTaskId` was no
longer recognized as a valid property as of v2.8.2

Because `ScheduleAction` is defined as a union of different types, we
must use [distributive conditional
types](https://www.typescriptlang.org/docs/handbook/2/conditional-types.html#distributive-conditional-types)
to capture the specific properties of the individual subtypes correctly.